YUK-6284 New customer account and device enrollment import process
- Fixed trying to delete the zero-th billing address.

+++ yukon-client/stars/src/com/cannontech/stars/dr/account/service/impl/AccountServiceImpl.java
@@ -378,7 +378,9 @@ public class AccountServiceImpl implements AccountService {
         /*
          * Delete billing address
          */
-        addressDao.remove(billingAddress);
+        if(billingAddress.getAddressID() != CtiUtilities.NONE_ZERO_ID) {
+            addressDao.remove(billingAddress);
+        }
         
         /*
          * Delete customer
